requirements.txt
chardet==3.0.4
defusedxml==0.5.0
Django==1.11
django-rest-framework==0.1.0
djangorestframework==3.8.2
idna==2.6
oauthlib==2.0.7
PyJWT==1.6.1
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
urllib3==1.22
